fix(db): alterar coluna id de SERIAL para BIGINT em tabelas existentes
- Alterado o tipo das colunas id de SERIAL para BIGINT para compatibilidade com Java Long.

+++ backend/src/main/resources/db/migration/V4__create_table_artista_album.sql
@@ -1,8 +1,8 @@
 CREATE TABLE artista_album
 (
-    id         SERIAL PRIMARY KEY,
-    artista_id INTEGER NOT NULL,
-    album_id   INTEGER NOT NULL
+    id         BIGSERIAL PRIMARY KEY,
+    artista_id BIGSERIAL NOT NULL,
+    album_id   BIGSERIAL NOT NULL
 );
 
 ALTER TABLE artista_album
